requirements.txt
transformers==4.52.1
torch>=2.0
sentencepiece~=0.2.0
nltk~=3.9.1
pandas~=1.5.3
protobuf>=3.20
nltk[data]

ollama~=0.5.1
sentence-transformers~=5.0.0
spacy~=3.8.7
h11~=0.14.0
ipython~=8.18.1
future~=0.18.3
pillow~=11.2.1
pip~=25.1.1
wheel~=0.38.1
filelock~=3.18.0
blis~=1.3.0
matplotlib~=3.9.4
regex~=2024.11.6
scipy~=1.13.1
scikit-learn~=1.6.1
click~=8.1.7
tqdm~=4.67.1
requests~=2.32.4
pyparsing~=3.2.0
pytz~=2025.2
rich~=13.9.4
pygments~=2.19.1
markdown-it-py~=3.0.0
PyYAML~=6.0.2
exceptiongroup~=1.3.0
sniffio~=1.3.1
cymem~=2.0.11
httpcore~=1.0.7
idna~=3.10
httpx~=0.28.0
mdurl~=0.1.2
charset-normalizer~=3.4.2
setuptools~=58.0.4
threadpoolctl~=3.6.0
mpmath~=1.3.0
sympy~=1.14.0
platformdirs~=4.3.6
thinc~=8.3.6
srsly~=2.5.1
wasabi~=1.1.3
weasel~=0.4.1
typer~=0.15.1
catalogue~=2.0.10
jinja2~=3.1.4
pydantic~=2.10.3
packaging~=24.2
langcodes~=3.5.0
preshed~=3.0.10
psutil~=6.1.1
tornado~=6.4.2
decorator~=5.1.1
murmurhash~=1.0.13
confection~=0.1.5
networkx~=3.2.1
importlib-metadata~=8.7.0
urllib3~=2.5.0
fsspec~=2025.7.0
shellingham~=1.5.4
MarkupSafe~=3.0.2
joblib~=1.5.1
anyio~=4.7.0
openpyxl~=3.1.5
SQLAlchemy~=2.0.36
python-dateutil~=2.9.0.post0
cloudpathlib~=0.21.1
certifi~=2024.8.30
six~=1.17.0
email-validator~=2.2.0
annotated-types~=0.7.0
zipp~=3.23.0
smart-open~=7.3.0.post1
wrapt~=1.17.2
tokenizers>=0.21,<0.22
safetensors~=0.5.3
spacy-legacy~=3.0.12
huggingface-hub~=0.33.4
fastapi~=0.115.6
starlette~=0.41.3
uvicorn~=0.32.1
marisa-trie~=1.2.1
language-data~=1.3.0
spacy-loggers~=1.0.5
rdflib~=7.1.1
textstat~=0.7.8
wikidata
beautifulsoup4
streamlit
syntok
streamlit-aggrid>=0.3.4
google-generativeai
SPARQLWrapper
faiss-cpu
ollama
sentence-transformers
scikit-learn
pandas
numpy
pyvis
jinja2